demo/requirements-demo.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
opensearch-py>=2.4.0
redis[hiredis]>=5.0.0
google-genai>=0.8.0
httpx>=0.27.0
python-dotenv>=1.0.0
numpy>=1.26.0
structlog>=24.0.0
